Refactor LUIS, QNA and Speech to use a single base component
