+++ Composer/packages/client/src/pages/botProject/RootBotExternalService.tsx
@@ -320,16 +320,16 @@ export const RootBotExternalService: React.FC<RootBotExternalServiceProps> = (pr
   const updateLuisSettings = (newLuisSettings) => {
     setSettings(projectId, {
       ...mergedSettings,
-      luis: { ...mergedSettings.luis, ...newLuisSettings },
+      luis: { ...mergedSettings.luis, authoringKey: newLuisSettings.key, authoringRegion: newLuisSettings.region },
     });
   };
 
   const updateQNASettings = (newQNASettings) => {
     setSettings(projectId, {
       ...mergedSettings,
-      qna: { ...mergedSettings.qna, ...newQNASettings },
+      qna: { ...mergedSettings.qna, subscriptionKey: newQNASettings.key },
     });
-    setQnASettings(projectId, newQNASettings.subscriptionKey);
+    setQnASettings(projectId, newQNASettings.key);
   };
 
   return (
@@ -418,7 +418,7 @@ export const RootBotExternalService: React.FC<RootBotExternalServiceProps> = (pr
         />
         <ManageLuis
           hidden={!displayManageLuis}
-          setDisplayManageLuis={setDisplayManageLuis}
+          setVisible={setDisplayManageLuis}
           onDismiss={() => {
             setDisplayManageLuis(false);
           }}
@@ -429,7 +429,7 @@ export const RootBotExternalService: React.FC<RootBotExternalServiceProps> = (pr
         />
         <ManageQNA
           hidden={!displayManageQNA}
-          setDisplayManageQna={setDisplayManageQNA}
+          setVisible={setDisplayManageQNA}
           onDismiss={() => {
             setDisplayManageQNA(false);
           }}
