+++ Composer/packages/client/src/pages/botProject/adapters/ABSChannels.tsx
@@ -369,11 +369,7 @@ export const ABSChannels: React.FC<RuntimeSettingsProps> = (props) => {
     };
   };
 
-  const toggleSpeechOn = async (
-    settings: { subscriptionKey: string; region: string },
-    isDefault: boolean,
-    attempt = 0
-  ) => {
+  const toggleSpeechOn = async (settings: { key: string; region: string }, isDefault: boolean, attempt = 0) => {
     setChannelStatus({
       ...channelStatus,
       [CHANNELS.SPEECH]: {
@@ -384,7 +380,7 @@ export const ABSChannels: React.FC<RuntimeSettingsProps> = (props) => {
 
     try {
       await createChannelService(CHANNELS.SPEECH, {
-        cognitiveServiceSubscriptionKey: settings.subscriptionKey,
+        cognitiveServiceSubscriptionKey: settings.key,
         cognitiveServiceRegion: settings.region,
         isDefaultBotForCogSvcAccount: isDefault,
       });
@@ -549,7 +545,7 @@ export const ABSChannels: React.FC<RuntimeSettingsProps> = (props) => {
       )}
       <ManageSpeech
         hidden={!showSpeechModal}
-        setVisibility={setShowSpeechModal}
+        setVisible={setShowSpeechModal}
         onDismiss={() => {
           setShowSpeechModal(false);
         }}
